move Panel.tsx into layout/ directory
